+++ diy_asylum_frontend/src/contentpages/namePage.js
@@ -31,5 +31,6 @@ const formElements = [
 export default makeContentPage({
   name: "Your name",
   formElements,
-  help: "Your complete name"
+  help: "Your complete name",
+  id: Symbol()
 });
